+++ scantpaper/app_window.py
@@ -122,6 +122,8 @@ class ApplicationWindow(
     _lockfd = None
     _pref_udt_cmbx = None
     _scan_udt_cmbx = None
+    _scan_udt_hbox = None
+    _scan_udt_button = None
     slist = None
 
     # Temp::File object for PDF to be emailed